Add NEW-withssl-buffered-3step.c
This shows how to insert a `BIO_f_buffer()` buffering filter BIO
in between the SSL and TCP-socket layers. Note that now when we
write to the SSL BIO, it writes to its underlying BIO, which is
the buffering BIO; if we want the write to actually flush all
the way to the TCP socket, we need to manually `BIO_flush()`
the SSL BIO after writing to it. We didn't need the flush in
"NEW-withssl-2step.c" because it was unbuffered.